For #24114 - Refactor HomeFragmentState into AppState
